Move package "loader" to package "service". - Load outside modules when Injector created. - Delete tests - Refactor
