--- orienteer-core/src/main/java/org/orienteer/core/loader/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author Vitaliy Gonchar
- * This package contains classes that load outside modules for Orienteer. (loader micro-framework).
- */
-package org.orienteer.core.loader;
