--- orienteer-core/src/main/java/org/orienteer/core/loader/util/ODependenciesNotResolvedException.java
@@ -1,10 +1,0 @@
-package org.orienteer.core.loader.util;
-
-/**
- * @author Vitaliy Gonchar
- */
-public class ODependenciesNotResolvedException extends Exception {
-    public ODependenciesNotResolvedException(String msg) {
-        super(msg);
-    }
-}
