--- orienteer-core/src/main/java/org/orienteer/core/loader/util/aether/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author Vitaliy Gonchar
- * Util classes for Eclipse Aether.
- */
-package org.orienteer.core.loader.util.aether;
