--- orienteer-core/src/main/java/org/orienteer/core/loader/util/metadata/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author Vitaliy Gonchar
- * Util classes for work with metadata.xml
- */
-package org.orienteer.core.loader.util.metadata;
