--- orienteer-core/src/main/java/org/orienteer/core/loader/util/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author Vitaliy Gonchar
- * Util classes for loader micro-framework.
- */
-package org.orienteer.core.loader.util;
