+++ orienteer-core/src/main/java/org/orienteer/core/service/OrienteerFilterInitModule.java
@@ -6,7 +6,7 @@ import com.google.inject.Singleton;
 import com.google.inject.servlet.ServletModule;
 import org.apache.wicket.guice.GuiceWebApplicationFactory;
 import org.apache.wicket.protocol.http.WicketFilter;
-import org.orienteer.core.loader.OLoaderStorage;
+import org.orienteer.core.service.loader.OLoaderStorage;
 
 import javax.servlet.FilterConfig;
 import javax.servlet.ServletContext;
