+++ orienteer-core/src/main/java/org/orienteer/core/service/OrienteerModule.java
@@ -114,5 +114,4 @@ public class OrienteerModule extends AbstractModule {
 		return version!=null?version:"";
 	}
 
-	
 }
