+++ orienteer-core/src/main/resources/log4j.properties
@@ -1,6 +1,6 @@
 log4j.appender.Stdout=org.apache.log4j.ConsoleAppender
 log4j.appender.Stdout.layout=org.apache.log4j.PatternLayout
-log4j.appender.Stdout.layout.conversionPattern=[%.5p] (%d) %.26c{1}: %m%n
+log4j.appender.Stdout.layout.conversionPattern=[%.6p] (%d) %.26c{1}: %m%n
 
 log4j.rootLogger=INFO,Stdout
